fix(failing build): fix build config files
- change env variable accesing methods for prod and dev envs

[Delivers #167401391]
