+++ build-utils/webpack.dev.js
@@ -1,13 +1,16 @@
 const { DefinePlugin } = require('webpack');
+const { config } = require('dotenv');
+
+const env = config().parsed;
+const envKeys = Object.keys(env).reduce((prev, next) => {
+  prev[`process.env.${next}`] = JSON.stringify(env[next]);
+  return prev;
+}, {});
 
 module.exports = {
   mode: 'development',
   devtool: 'eval-source-map',
   plugins: [
-    new DefinePlugin({
-      'process.env': {
-        NODE_ENV: JSON.stringify('development'),
-      },
-    }),
+    new DefinePlugin(envKeys),
   ],
 };
