+++ build-utils/webpack.prod.js
@@ -7,6 +7,7 @@ module.exports = {
     new DefinePlugin({
       'process.env': {
         NODE_ENV: JSON.stringify('production'),
+        API_URL: process.env.API_URL,
       },
     }),
   ],
